requirements.txt
mss
Pillow
winrt
